Add 'diagnose datadog-connectivity' subcommand (#11913)
* feat(diagnose): add empty new diagnose subcommand 'datadog-connectivity'

Create a new 'datadog-connectivity' which will be used to diagnose connectivity issues
between the Agent and Datadog endpoints. This command is empty for now, but will be implemented
in futures PRs.

There is also a new subcommand 'metadata-availability' which is doing the same job as the 'diagnose'
command. The 'diagnose' command might evolve in the future to run both 'datadog-connectivity' and
'metadata-availability' checks.

Signed-off-by: Nicolas Guerguadj <nicolas.guerguadj@epita.fr>

Co-authored-by: Vickenty Fesunov <vickenty@users.noreply.github.com>