Attempt to reduce the amount of boilerplate needed to add a product
At the cost of having to manual manage some type casting, attempt to
reduce the amount of boilerplate needed to add a new product.

Adding a new product should now only consist of:
    - Adding the new product string and adding to the products list
    - Defining the config structure and a function to parse it
    - Adding that parse function to parseConfig in respository.go
    - Exposing a way to retrieve typed configs

Note that this also changes the concept of an "updateResult". The RFC
suggests providing updates to products that explicitly list the "new",
"updated", and "removed" configs...but both APM and LD don't care about
this and just want the current state of configs when updates come in. As
a result, this implementation does not do the more advanced reporting at
the moment.