Only store updated roots if everything in the update validates
We don't want to be in a situation where we apply the root updates, but
there's an issue with something else in the update that causes our state
to be partially updated.

To accomplish this we use a temporary root TUF client that is a clone of
the current one. We then apply the root update to the temporary client,
and at the end if everything checks out, we make the temporary client
the new official client.