[RCM] Properly filter by product the list of relevant configs to clients
The ClientConfigs field in the ClientGetConfigsResponse struct is
supposed to contain a list of config paths that correspond to all of the
config files a client should maintain at a given point in time.

At the moment, it's not respecting the client's wishes and only
providing config files for products that the client supports. This is
not an issue with the current Go client contained in this agent release,
but this will be important for other clients that will need to work with
this release of the agent.