Modify AC.GetAllConfigs to wait and then return all loaded configs
Previously, this call would essentially invoke LoadAndRun itself, which
results in duplicate polling of pollable config providers, and does
nothing for non-polling providers.  Also, when using the more accurate
reconcilingConfigManager, it would only return configs that were newly
discovered in that poll operation, not all configs.

With this change, the function is simpler:
 - wait until a separately-invoked LoadAndRun completes
 - map over all configs and return them

This does, however, require users of the AC.GetAllConfigs function to
invoke LoadAndRun, but that should be a normal part of startup anyway.