[RCM] Add a method to pull only the TUF version state
Pulling the full state is a non-trivial operation, and we were doing it
on every update request from an agent/tracer client. These clients ask
for updates quite frequently and this was causing the agent to use
more CPU than intended for remote configuration updates.

To fix this, we split out full state checks (which ARE needed for the
agent's CLI for users to query the remote config state) from TUF Version
state checks, which is all that is needed for trace/agent client updates
as well as getting updates from the remote configuration backend.