Update the security-agent to use the new RC client
Update the security-agent to use the new RFC compliant remote
configuration client.

The internal changes to the remote config client do not impact the
content being sent to the security-agent. The main change is the means
with which the security-agent receives the updated configs. Instead of
asking the client for a channel on which config updates are sent, a
callback function is registered that the client will call whenever there
is an update.