Update the trace-agent to use the new RC client
Updates the trace-agent's use of the remote config client to work with
the new RFC compliant client.

The internal changes to the client do not impact the content being sent
to the trace-agent. The main change is that instead of asking the RC
client for a channel on which updates will be sent, a callback function
is registered that the client will call when updates involving APM
Sampling occur.