pkg/obfuscate: add support for PostgreSQL JSON operators (#12026)
Add support for PostgreSQL specific JSON and JSONB operators to the obfuscate package.  This commit adds special handling in the tokenizer for multi-character json operators (#>, #>>, #-, @>, <@, ?|, ?&, ->, ->>).  These operators will only be parsed when DBMSPostgres is passed as the DBMSType in the SQLConfig.  This fixes an issue where operators with a `#` were incorrectly interpreted as a comment.