[RCM] Fix storage of backend state in client (#12530)
There was a bug in the communication between the backend and the agent
involving backend state that is supposed to be opaque to the agent and
its clients. (allowing the backend to store whatever it wants downstream
to help optimize the update process or recreate state if needed)

In order to fix the bug, we needed to introduce a new field and change
the data type from json.RawMessage to []byte.