fix(remote-config): Throw TUF errors when getting configs from Director
Currently any TUF error happening when getting a config from the
Director results in a Not Found error. If our snapshot is expired or
simply does not exist, we want to be able to get a meaningful error.
If that change is not made, we may fall into a case where we don't know
what is happening when we have a TUF issue in the agent or the backend.