[remote tagger] Don't retry stream when shutting down
We noticed the following error in production, happening multiple times
for the same container:

```
error received trying to start stream: tagger stream not started
```

This happens because the remote tagger tries to reconnect when it should
actually just start shutting down. This commit prevents these retries
from happening, by trying to re-establish a stream at the beginning of a
loop, instead of the end, so it'll only fail once, and not retry.