[criutil] Prevent gRPC connection leak on errors
The context passed to `grpc.DialContext` does not actually close the
connection, and instead just stops the dialing process itself. To
actually close the connection, one still needs to call `conn.Close()`.

`init` is supposed to be called just once, but it gets retried on
errors. This means that we start to accumulate unused gRPC connections
until the operation finally succeeds. To prevent this leakage, we now
close the connection on errors.