pkg/trace/stats: consider http.status_code from Metrics
This change makes sure that stats computation consider the
http.status_code aggregation key when present in Metrics too, and even
prioritise it. The reasoning is that the status code is an integer by
nature and tends to get sent as such. This should not cause a defect in
the generation of APM stats aggregation keys and tags.