cca_in_ad -> replace kubernetes/docker launchers with container launcher (#12440)
* cca_in_ad -> replace kubernetes/docker launchers with container launcher

The kubernetes and docker launchers have overlapping functionality, and
both have serious bugs and missing functionality.  The container
launcher implements the same behavior, with fewer bugs and more
functionality.  In particular, it

 - handles config from files (conf.d) for containers under kubernetes
 - handles changes to pod annotations
 - correctly logs from file or socket based on `logs_config.{k8s,docker}_container_use_file`
 - captures logs under containerd when `k8s_container_use_file` is false

The entire container launcher is hidden behind the 'docker' build tag,
which really means "containers" and not specifically docker.

Outside of the situations where they have obvious bugs, this launcher
attempts to preserve behavior of the kubernetes and docker launchers in
the details, such as defaults for the "Source" and "Service" properties.
It does not preserve those launchers' behavior with respect to `agent
status` output, as that behavior was not very intuitive and sometimes
incorrect.

In this commit, the container launcher includes its own, private notion
of a "tailer" which wraps existing logs-agent functionality: for
sockets, it wraps the socket tailer; for files, it wraps file _sources_.
Both of these wrappers are ugly and awful but allow reuse of the
existing functionality without modifying it.

Co-authored-by: Brian Floersch <brian.floersch@datadoghq.com>