Benchmarking added for kernel header downloading (#12684)
* benchmark added for ubuntu header downloading

* bench marking for debian header download added

* linter errors fixed

* Using the TempDir function to create temporary directories, when appropriate

* Only adding correct repository for Ubuntu.

Aptly uses leveldb for storing remote repositories. We search for the
required pacakge in these repos. The order in which leveldb returns the
repo while iterating is not deterministic. This adds variability in the
benchmarks, based on how long it takes us to find the correct repo. To
prevent this, and stabilise the results, this commit only adds a single
repository to the Ubuntu backend for downloading kernel headers.

* fix linter errors