feat(remoteconfig): Bypass cache on new client request
Make sure that every client that connects for the first time to the
agent triggers a cache bypass, ie a request to the backend. This is
to avoid cold starts where the client has to wait one minute to get
its configuration.
To avoid handling too many requests in the backend, we introduce a
rate limiter for this specific feature.