Start tagger before `api.StartServer`
This addresses a race condition affecting the remote tagger. As
`tagger.SetDefaultTagger` happens inside `common.LoadComponents`, before
that happens the agent is running an empty tagger that will never see
any updates. Since there's a possibility that remote taggers will start
between `api.StartServer` and `common.LoadComponents`, there's a race
condition that essentially breaks the remote tagger completely, as we've
seen with missing tags for both the trace and process agent.

To prevent this race condition, this commit moves the initialization of
the tagger *before* starting the agent's HTTP API.