[tagger] Prevent remote tagger wipe on keep-alive
This prevents a keep-alive sent from the core agent (or any empty
message, for that matter) from wiping the remote tagger's contents if it
was previously in an unready state. Although this would be quite rare
and short lived, we suspect this might've been contributing to missing
tags in data being reported from the process agent after the
introduction of the keep-alive messages.