pkg/obfuscate: update dgraph-io/ristretto to outcaste-io/ristretto (#12559)
The dgraph version of the library is abandoned and the outcaste-io one
is now the official fork. The new version removes a dependency on the
glog library, that can cause conflicts due to global initialization, see https://github.com/outcaste-io/ristretto/pull/3.

The same glog library has caused issues in this repo in the past as well https://github.com/DataDog/datadog-agent/pull/1310.