[remote tagger] Always honor `<-ctx.Done()`
In some rare instances where the remote tagger is stopped at the same
time as either the health check or telemetry ticker trigger, the
`select` would not see that `t.Ctx` was already cancelled, and would
carry on trying to re-start the stream. This messes with the shut down
of other agent components, as `Tagger.Stop()` seems to block in that
case.

To make sure that `<-t.ctx.Done()` is always evaluated, we now re-start
the loop on the health check and telemetry ticks.