Copy conn_tuple_t to stack in the subprogram for tls parsing.
The function proto for bpf_map_update_elem defines the second argument
to be of type ARG_PTR_TO_MAP_KEY, which is a pointer on stack of the map
key to be used. On most kernel, we can actually provide a pointer to a
map value as well. However on kernel 4.4 the verifier does not allow
this. This commit copies conn_tuple_t, which is used as the key
in the bpf_map_update_elem call, to the stack.