Optimizing read_into_buffer helper (#12617)
* reading complete size in fragment buffer, and post processing later

* code restructuring

* post processing when batch of transactions is accessed

* renaming and cleanup

* updating fragment_len in all places

* fix verifier error

Caluclating read_len earlier in the function and returning it later
causes the compiler to save the value on to the stack, causing the
verifier on older kernels to loose the state information.
http->fragment_len is then set by offseting into the stack and setting
it to read_len.
This causes the verifier to complain about arithmetic between fp and a
register with an unbounded value.
Directly returning the read_len keeps the value in the register, and
preserves the state information.

* reading HTTP_BUFFER_SIZE in 'read_into_buffer_skb' function

* revert 'read_into_buffer_skb' change

* fixed read_into_buffer_skb after last revert

* comments added

* attaching a stub program for http socket filter
This is done to bypass memory limitations imposed by the kernel when
attaching a filter to a socket. Using a stub program which branches into
a subprogram allows us to avoid the check which enforces this limit.
See for description of the problem:
https://datadoghq.atlassian.net/wiki/spaces/NET/pages/2326855913/HTTP#Program-size-limit-for-socket-filters
The check enforces that there is enough space in the sockets option
memory buffer:
https://elixir.bootlin.com/linux/v5.4/source/net/core/filter.c#L1198

* HTTP_BUFFER_SIZE increased

* BLOCK_SIZE changed back to original value

* revert HTTP_BUFFER_SIZE

* renamed http socket filter entry point

* Remove passing of fragment_len to userspace

It was noted that it is unnecessary to pass the length of the fragment
buffer to userspace, because the userspace parsing of the HTTP path will
ignore any junk data present in the fragment buffer. This commit removes
the bulk of the changes which were introduced in this PR.

* comments added

* add runtime asset file