Using MaxTrackedConnections as the size of filter_args
It was falsely assumed that the __cgroup_bpf_run_filter_skb and the
socket filter execution was synchronous. As such the size of filter_args
was set to #cpus. This turned out not to be the case and it was noted
that we were failing to the the filter args in the socket filter
sometimes.
Would be useful to add telemetry here to see when filter_args gets full,
or when we fail to get the argument in the socket filter.