[RCM] Correct runtime ID and client ID usage in TracerPredicate protobuf
Within the agent, we were referring to field 1 of the TracerPredicate
struct as runtimeID when the remote config backend was actually using
this field to refer to the clientID.

We still want to be able to check against runtimeID. To fix this
we'll rename runtimeID to clientID to match the backend, add a new check
against the ClientID specifically so that our predicate evalution works
properly with what the backend is setting up for us, and add a new field
to the protobuf defintiion to allow us to still have predicate against
runtimeID should that be desired by users of the system.

The end result is that the RC backend is now in sync with the
agent, both semantically and with field names.