[ad] Wait for first workloadmeta bundle on provider start
This fixes an issue in `agent check`, where a subtle change in timing
introduced in #12968 meant that configs did not get scheduled in time
for `agent check` to see them, effectively appearing as if they were not
present. This change ensures that `kubelet` and `container` providers
only return from `Collect` if they've received at least one event bundle
from workloadmeta.

This also changes workloadmeta to actually allow the sending of empty
bundles, as that's necessary to signal that the subscription has started
and there are indeed no entities in the store. Otherwise, config
providers in non-containerized environments will hang indefinitely.