[ad] Generate configs only once for k8s containers
Previously, we'd generate configs for the container events themselves,
and again when handling the pod event. When the container belonged to an
AD-annotated pod, it would briefly have a `container_collect_all` config
before honoring the one in the annotations. Now, since the configs are
generated entirely at once, AutoDiscovery can ignore the
`container_collect_all` correctly the first time.