Change fingerprinting functions because of debian 10 on arm64
debian 10 on arm64 does not expose the various *_ops structures we were
previously using to fingerprint socket types. Now we use the functions
pointer to recognize the type of socket.

Moreover, we no longer checkout for tcp_prot structs, because inet
sockets will always have the inet_stream_ops as the function table for struct socket* objects. The tcp
specific functions are pointed to by the functions table in struct
sock*.