[omnibus] Vendor libcrypt.so.2 in the Agent package (#13135)
Adds a new software definition, libxcrypt, that builds the libcrypt.so.2 library on Linux. This library is a dependency of Python.

Newer OS versions may not provide libcrypt.so.1 in their system, so we need to bundle it in the Agent to make sure libraries and binaries (mainly the Python interpreters & libraries) using it can function.

And since we are building & bundling a libcrypt library ourselves, we might as well build libcrypt.so.2, a more modern version of the library.