Make the test case more robust
Using distinct origin ids in the test makes createOriginTagMaps
concurrently write to the hash map, triggering a runtime crash, rather
than just a warning with -race.

Sync up all goroutines calls to increase the probability of a fatal
collision.

Waiting for goroutines to finish makes sure that any crash happens
while the test process is still running (if the test is the last in
the queue, the only test running).