Avoid sending inventory data for unscheduled check
In a k8s environment the container linked to a check instance might
stop/be killed while the check is running. Since we don't have a way to
stop a Python thread we unschedule the check but let it finish its run.
This means that inventories might receive metadata for a check that no
longer exists in the collector and will never be cleaned.

This commit cleans the check metadata cache every time we generate the
payload. This means we no longer send information about checks not in
the collector. Therefore, the `check` command has to manually print
metadata instead of relying on the status page.