WIP. Begin work on remove hard-coded build path. Hard-coded build path is an artifact of previous Go versions, where datadog-agent had to be within GOPATH.  It no longer needs to be.
This change should be a <noop>, but opens opportunities for mounting
different directories in a build container for ongoing development.

Fix some places where merge conflicts were resolved incorrectly.

Fix some places where the execution would occur on the mounted
filesystem rather than in the container.

Fix using global environment variables when local script variables were
sufficient.

Update lint scripts to also not require hard-coded directories.

change `lint.ps1` to not use environment variable for variable not needed outside
of script.

fix expansion of modcache files

By popular demand, delete the modcache tar files even though they're in
the container.

Hopefully final feature creep.
Check to see if GOMODCACHE is set.  If not, don't bother
with the expansion because it will take up a lot of time and space
for nothing.

And then, use that variable as the target for the unzip rather than
unzip to hard-coded location.  If GOMODCACHE was pointed elsewhere,
the expansion would also be wasted.