Add the possibility to change at runtime some activity dumps parameters: max_dump_size, cgroup_dump_timeout, rate_limiter and traced_event_types.
For the first one, the change is imediate.
For the others, on next load controler period they will trigger the send of the current dumps and a fresh start of new partial ones with the new config.

Other change: all probes needed for all possible AD traced events are now loaded at startup (to be able to add them later on).