[collector] all tests using mock senders (i.e. checks) does not enable the no agg pipeline.
Some of them are modifying the internal buffer of the shared logger or some configuration stuff. While doing so, the unit test may also be initing the demultliplexer/aggregator causing races. In order to avoid this, all tests using the mock Sender (used with Checks collector) are not enabling the no-aggregation pipeline.