pkg/trace/api: bring back containerID detection through UDS for cgroup2 (#13224)
This PR brings back cgroup v2 support for incoming trace information.
(Support had to be reverted as the implementation created a reference
from the trace-agent to the datadog-agent which caused issues).

This allows the trace-agent to look up container IDs for incoming traces
with the cgroups v2 API (and the PID of the source of the traces which
can be had by sending traces over UDS).

Cgroups v2 api has been out for a long time at this point, this will
allow customers to continue to get container IDs as they upgrade to
systems that only make use of the v2 api.

Unfortunately some of the code here is duplicated from the collector code
(which was what the previous implementation called but had to be reverted
due to its dependence on datadog-agent). All duplicated code has been
marked with comments and represents a small fraction of the overall code
here.