pkg/trace: add endpoint to proxy statsd over HTTP (#13393)
This PR allows DogstatsD messages to be proxied through the Trace Agent over HTTP, using the same port or Unix domain socket as tracing. This reduces the number of endpoints required to be configured in order to use Datadog APM with both traces and runtime metrics.

This change also adds a new field to the Trace Agent Config which indicates whether 'use_dogstatsd' is true or false. This could use a few more tests before merging, and perhaps it isn't strictly necessary. Feedback on this welcome.

The tests include:

unit tests that confirm the correct status code based on the provided configuration
an end-to-end test which spins up a DogStatsD server, and sends an HTTP request to the proxy, verifying that the data shows up correctly in DogStatsD.