[external metrics] Always choose autoscaling >= v2beta1
In Kubernetes <1.23, the preferred version for HPA seems to be v1 (which
we have never supported) instead of any of v2beta1, v2beta2, v2. To
force >=v2beta1 support, we need to ignore that older version, and
instead look through all versions supported by the API server and choose
the highest one.