serverless: appsec monitoring poc
POC showcasing the integration of the ASM threat monitoring into the universal instrumentation API.

To do so, the start-invocation endpoint is used to monitor the lamdba invocation and event payload.
The HTTP-based events, such as the API Gateway event, can then be monitored by the WAF, whose security
events are added to the inferred span tags.

Note that the appsec source code was copied from dd-trace-go but the plan is rather to share it.