pkg/{autodiscovery,config}: Adjust init and APIs to avoid dep cycle.
There was a circular dependency between pkg/config and
pkg/autodiscovery/common/types that was resolved by having types set a
global value in pkg/config on init(). However, this means that a user of
pkg/config cannot rely on the prometheus functionality without also
importing pkg/autodiscovery/common/types to get the init() function to
run. If it is not imported, nil values are unexpectedly present, and
panics occur.

This is not a good way to resolve a circular dependency. Instead, the
code has been reorganized so that pkg/autodiscovery/common/types does not
depend on pkg/config, and instead is parameterized (config values are
passed in). This allows pkg/config to import
pkg/autodiscovery/common/types and avoid dealing with init() at all,
since it uses functionality from that package.

Fixes #13595