Extract common static parameters to a struct
Simplify signature of functions in enrich.go by extracting common
static parameters to a struct, allowing us to pass the whole bundle as
a unit. This helps separate sample data from configuration (which
previously was interleaved in some functions) and make future work
easier, as less code and tests need to be updated.