[workloadmeta] Extract EntityMeta in kubemetadata
This makes sure that KubernetesPod.EntityMeta is always present. This
information being missing caused a problem in the kubecontainer
autodiscovery provider, since namespace/name is used as a cache key.
When running this collector concurrently with the kubelet one, it would
often happen that the kubelet would unset a deleted pod before
kubemetadata would do the same, and ultimately the
workloadmeta.EventTypeUnset event would not contain EntityMeta anymore.