[RCM] Support filtering expired configs in agent
Adds the ability for the backend to indicate the expiration time for a
given configuration ID. If the config is expired, the agent will not
deliver that configuration to downstream clients. This allows config
that might be time sensitive to expire even if the connection to the
Datadog backend is not active to receive updates.