[RCM] Allow the backend to override the agent's refresh interval (#13660)
Certain products want the agent to refresh at a more frequent interval
in order to ensure config updates propagate in a timely manner. In order
to accommodate this, we'll allow the backend to send refresh interval
hints to the agent via the TUF director targets custom field. If the
agent wasn't given a manual refresh interval by a user (this always
takes priority) the agent will adjust its default refresh interval to
the backend provided value, assuming it is in valid range.