[RCM] Add an unverified repository and client for remote config
Ideally the agent clients perform some basic TUF verification of their
own, however, since the core agent is doing full verification, if we
trust the connection between the clients and the agent we can forego
verification. This simplifies things on the client side, as they don't
need to have the base root file embedded into their binary in order to
create a Repository.

This is primarily for the Go tracer, which utilizes the Repository
structure for storage of its state as well. Until root rotations are
fully off the ground, this prevents us from having to publish a new root
file and requiring end users to either reconfigure or rebuild their
applications if we need to publish a new root in the interim.