Keep senders until check stops running
When a check is unscheduled while it is still executing, previously
Cancel would destroy the sender for the check, but does not abort
check execution. If a check has configured sender (e.g. to omit
default hostname), such configuration would be lost when the check is
cancelled. This problem existed for both go and python checks.

This patch postpones DestroySender call until after the check has
finished executing. A wrapper is inserted around every scheduled
check. When the check is cancelled, the wrapper spawns a goroutine
that waits for any outstanding invocations of Run to finish, and then
destroys the check sender.