[Logs - Journald] Allow multiple tailers on the same Journal (#13874)
This change allows multiple tailers to read the same journal. This is particularly useful when users want to apply different tags to different units from the same journal. This change also introduces the config_id setting to the journald tailer config. The config_id is used to uniquely identify the configuration so the agent can pick up where it left off in the event of a restart.