Extract `agent check` subcommand into `pkg/cli/subcommands`
This subcommand is used by more than just the `agent` cmd, which is a
common situation for other subcommands that need to be available in
other binaries as well.

This commit extracts the `*cobra.Command` returned by
`cmd/agent/subcommands/check.Commands` into a separate package that can
be called by other binaries, namely `cluster-agent` (to be done in a
separate PR).

There is one downside to this change, in that it requires
`cmd/internal/standalone` to be made non-internal, at least for the
moment. We may decide to move everyone that depends on it to `pkg/cli`
too, or break `standalone` into more granular packages, but that's out
of the scope of this commit for now.