pkg/trace/api: add /dogstatsd/v2/proxy that splits on newlines (#13857)
What does this PR do?
Adds a v2 endpoint for the DogStatsD proxy through the Trace Agent which splits payloads by newline. The /dogstatsd/v1/proxy endpoint has also been changed to parse requests in this way, since the risk is negligible that someone is relying on the previous behavior.

Motivation
Previously, clients would have to send a request for every single stats payload that they wanted to send over to the proxy. Now, they can send a large request with several stats requests separated by newlines to this endpoint, lowering the load on the tracer side.

An alternative would be to simply change the v1 endpoint to separate by newlines and not create a v2 endpoint. We can't do this since a current Agent release contains this v1 endpoint. The issue would arise if the NodeJS tracer is released with their logic hitting the v1 endpoint. There would be no way for the NodeJS tracer to guarantee that a customer is using the "newer" version of the v1 endpoint in the agent, instead of the "older" version that doesn't split by newlines. In that case, it could be sending over large payloads separated by newlines that are not currently split or processed in the Core Agent, and the status code of OK from this endpoint would make it look like everything was working as expected. Creating a v2 endpoint and having the NodeJS tracer send payloads over to this prevents this versioning issue in the agent.

This builds on top of #13393

Possible Drawbacks / Trade-offs
If a customer is currently relying on the /dogstatsd/v1/proxy endpoint not splitting on newlines, this will be a breaking change for them. But this is an endpoint that would only ever be used by Datadog's tracers, which none of them are today, so no customers should be affected by this.

Describe how to test/QA your changes
There are unit tests covering this use case, and there are manual tests that can run this end-to-end with the NodeJS tracer.