protocol-classification: Added HTTP/2 tests and fixed classification problems with NAT and streams
When dealing with binary stream protocol like HTTP/2 it might be impossible to classify the protocol both on the request and response
In HTTP/2 we can only identify the protocol by the first 24 bytes of the request.

When NAT is involved (like SNAT) there is a mismatch between the address the server is bind to (server address) and the address the client is communicating with (target address)
thus we don't have classified protocol for the response.

The solution is to save the protocols for both the sock connection tupple and the inverse skb connection tuple.

Added a UT to test gRPC (HTTP/2) behavior