pkg/trace: remove the appsecevts intake proxy
The first versions of ASM were using their own event intake while they now
rely on the APM trace intake.

It has been deprecated over the past quarters and it is now totally
disabled and removed from Datadog's backend. It is safe to remove this code
from the trace-agent that is no longer used by anyone.