Keep the check sampler alive until it is flushed
When a check is deregistered, we used to immediately delete the check
sampler, dropping all metrics sent by the check since the last flush.

To reliably send metrics from a check we need to keep the sampler as
long as there is potential for it to receive metrics, and then again
until it is flushed.

This patch adds a new message type, deregisterSampler, sent via the
metrics channel, to indicate that the sampler will not receive any
more metrics. Due to FIFO nature of the channel, it will be handled
after all the metrics sent by the check. At this point we know no more
metrics will arrive and can drop the sampler once we flush the
accumulated metrics. This is achieved with a flag, which is raised
when the aggregator processes deregisterSampler message, and the flag
is checked during flush: if the flag is set we know that no more
metrics will arrive for this particular check sampler.