Fix cluster check metrics sent with host tags (#13925)
* Keep senders until check stops running

When a check is unscheduled while it is still executing, previously
Cancel would destroy the sender for the check, but does not abort
check execution. If a check has configured sender (e.g. to omit
default hostname), such configuration would be lost when the check is
cancelled. This problem existed for both go and python checks.

This patch postpones DestroySender call until after the check has
finished executing. A wrapper is inserted around every scheduled
check. When the check is cancelled, the wrapper spawns a goroutine
that waits for any outstanding invocations of Run to finish, and then
destroys the check sender.

* Apply check wrapper in the collector

Collector is the one calling Check#Cancel, which needs to go through
the wrapper as well.

* Move check wrapper to a better location

Since it is used internally by the collector, make it an internal
subpackage of the collector.

* Remove CheckBase#CommonCancel

Sender lifetime is now managed centrally by the collector, so
individual checks no longer need to destroy the sender manually.

* Add releasenote