Send metrics from a de-scheduled check (#14056)
* Move telemetry updates into handler methods

Make a single method responsible for everything, so we can refactor
the channel code without impacting functionality.

* Send check metrics and sketches via the same channel

Make samples, sketches and commit messages be handled by the
aggregator in the same linear order they were sent by the sender.

This fixes a small race where sketches sent before Commit() could be
handled after the Commit() was applied to the sampler (see
sender_test.go change around 467), delaying the sketch delivery
slightly.

* Keep the check sampler alive until it is flushed

When a check is deregistered, we used to immediately delete the check
sampler, dropping all metrics sent by the check since the last flush.

To reliably send metrics from a check we need to keep the sampler as
long as there is potential for it to receive metrics, and then again
until it is flushed.

This patch adds a new message type, deregisterSampler, sent via the
metrics channel, to indicate that the sampler will not receive any
more metrics. Due to FIFO nature of the channel, it will be handled
after all the metrics sent by the check. At this point we know no more
metrics will arrive and can drop the sampler once we flush the
accumulated metrics. This is achieved with a flag, which is raised
when the aggregator processes deregisterSampler message, and the flag
is checked during flush: if the flag is set we know that no more
metrics will arrive for this particular check sampler.

* Fix race in the aggregator test

IsInputQueueEmpty() is not sufficient, as the message may have
received from the queue, but did not have the effect we want. Instead
of checking the channel, wait for the flag change directly.

* Fix data race in the test

This test now runs in parallel with the main aggregator thread, so we
need to use locks when poking at the aggregator internals.

* Move code to improve incapsulation

* Make test utility method available for tests only

* Update collector tests

Update tests for cancelled checks to verify that metrics sent during
or immediately before the cancel are sent correctly.