pkg/collector/corechecks/cluster/helm: added a new tag: helm_revision_is_most_recent
AFAIK it is not possible to sort metrics per tags/labels. At the moment it is therefore difficult to be able to query which one is representative of the most recent release.

This change aims to ease the identification of the most recent releases' revisions amongst the reported dataset. Introducing a new `helm_revision_is_most_recent` tag to metrics and events, this should give us some further filtering capabilities.