Fix metric loss with re-scheduled checks (#14228)
* Fix metric loss with re-scheduled checks

Since delaying sampler destruction for cancelled checks, checks that
were quickly re-scheduled were getting errors from GetSender because
the sampler for the check was still in the aggregator. A bit later,
once deregisterSampler message is handled and the sampler is removed
during flush, metrics from such checks would be completely lost.

This patch fixes the issue by making sampler re-use a normal condition
instead of an error: now that the sampler can outlive the check (this
is because we want to send the metrics from the previous run still
waiting for the flush) it isn't an error to have a sampler exist when
the check is scheduled.

To prevent race condition when registerSender may be called before
deregisterSampler message is even processed by the aggregator, sender
creation (or reuse) is now handled by a dedicated message as well. As
long as registerSender is called before the check run, sampler will
be created before metrics arrive: both sampler creation message and
metrics are sent via the same channel.

* Adjust tests for delayed sampler creation

* Separate test cases to avoid data races on the globals

Tests that re-init the demultiplexer conflict with the tests that do
not, because background goroutines for checks access the demultiplexer
instance when the next test starts.

This isn't an issue during normal agent run, because then the global
demultiplexer instance is read-only and not modified past initialization.

* Stop the collector before demux

Prevent the race in the tests where test workers access global demux
instance at the same time as demux.Stop clears it.