[RCM] Add the core remote config service and a client to the cluster-agent
Adds the core remote config service to the cluster-agent, as well as an
instance of an "agent client" that communicates directly to the core
service instead of over a secure GRPC connection.

This serves as a starting point for any team wanting to build out RC
functionality in the cluster-agent.