[Serverless] Add Azure App Service metadata to root spans (#14285)
If DD_AZURE_APP_SERVICES is enabled in the environment, the trace agent
will now add Azure App Service metadata to the root span of the trace.

The metadata is required for proper billing and view in the app. This
approach supports all runtimes, rather than updating each tracer
independently.