pkg/trace/watchdog: retrieve PID from same procfs that psutils uses (#14273) (#14308)
* pkg/trace/watchdog: retrieve PID from same procfs that psutils uses

The agent has historically used os.Getpid to get its PID. It then uses
this PID to retrieve its cpu/memory/etc. stats from gopsutil. Gopsutil in
turn looks in a procfs either configured by the HOST_PROC env variable or
/proc by default to find the process's stats.

This is a problem when we run the trace agent in a container and mount
the host's procfs in the container at /host/proc. When /host/proc is
mounted, the trace agent automatically sets HOST_PROC in pkg/config,
presumably for the process agent.

The trace agent looks up its PID with os.Getpid, which returns the pid
for the container's namespace, and then uses gopsutils, which is using
the host's PID namespace to get the stats.

This results in either missing stats when gopsutils can't find the PID,
or wrong stats when the PID corresponds to a different process in the
host's PID namespace.

Instead, we will now look up our PID when retrieving stats by reading the
'self' file from the same procfs that gopsutils will use.

Fixes #8068

* cleanup

Co-authored-by: Kyle Nusbaum <kyle@datadog.com>