Add hidden cli args to set paths for pythons 2 and 3
This can be helpful for working and testing things with local agent
builds. It's similar to the existing use-sys-python, but it lets us
simulate agent 6 scenarios without going through an omnibus build.

Ideally we shouldn't need this and the logic to find the python
executable to work with would be shared with the collector's code, but
that would require much more work.