[single-machine-performance] Push agent containers to SMP ECR
This commit is an attempt to introduce pushing containers from Agent CI for
single-machine-performance's Regression Detector in our isolated
infrastructure. Much like we have done for vectordotdev/vector we intend to run
the Regression Detector on Agent changes, giving a reasonable statistical
guarantee that a change does or does not modify Agent performance by more than
random chance. In order for the Regression Detector to run jobs it must have
access to a 'baseline' and 'comparison' target. Baseline in this project would
be a container built from current `main` branch, comparison would be a container
built from the tip of a PR.

The main thing demonstrated here is that the team credentials SMP has created
for Agent are functional and are able to push up a containers. This PR is,
however, a temporary thing and the method in #14438 is preferred once we have
all the correct cross-account permissions in place. This PR will unblock
single-machine-performance team meanwhile. Contrary to the approach taken in
PR #14438 I have duplicate the agent7 build job and adjusted the ECR/tag
location, logging in with credentials we have stored in project secrets. This is
not aesthetically pleasing nor resource efficient but we expect this situation
to last until at most early December.

In a later commit we will introduce job submission and will rely on being able
to compute the tag of a previous pipeline's container from available Gitlab
metadata, specificall `CI_COMMIT_SHA` for the comparison container and whatever
metadata maps to the base branch's current SHA,
`CI_MERGE_REQUEST_SOURCE_BRANCH_SHA`?

REF SMP-208

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>