Extract the flare component to its own bundle
"pkg/flare" package causes a 30MB increase in binary size when linked
with. Adding the flare component to the core bundle increases the size
of binaries using the core bundle but weren't compiled with flare support..

Since the bundle params are not defined at the bundle level we can't
include specific components from a bundle without linking to the entire
bundle. For this reason "flare" is in its own bundle, for now.

Once "pkg/flare" has been fully migrated to the component or the parameters
for components are at the component level we will merge back "flare"
into the "core" bundle.