[RCM] Explicitly use our fork of go-tuf
We were using our fork of go-tuf within the agent, but only as a replace
directive. This worked fine for the core agent code, but there are agent
packages that are themselves modules that can be imported externally
that relied on agent code behind this replace directive. In addition,
we were accidentally not using our fork in the `pkg/remoteconfig/state`
module.

If external code relied on any of these packages, or something that
itself relied on those packages (dd-trace-go), if a user ran `go get -u`
they would end up having go-tuf be updated to the most recent upstream
version (v0.5.1) which is incompatible with our code, as go get -u does
not treat v0 modules as special and assumes minor versions are safe,
even though semantic versioning does not provide such a guarantee.

To fix this we want all code explicitly requiring our fork. We modified
our fork to present itself as its own unique module path. This change
updates all references within the agent to explicitly refer to our fork
and not rely on replace directive magic.